Move finish function out of BrowserSwitchClient and into BrowserSwitchPendingState.
