+++ CorePayments/src/main/java/com/paypal/android/corepayments/browserswitch/BrowserSwitchPendingState.kt
@@ -1,5 +1,6 @@
 package com.paypal.android.corepayments.browserswitch
 
+import android.content.Intent
 import android.util.Base64
 import androidx.core.net.toUri
 import org.json.JSONObject
@@ -23,6 +24,35 @@ data class BrowserSwitchPendingState(val originalOptions: BrowserSwitchOptions) 
         val jsonBytes: ByteArray? = json.toString().toByteArray(StandardCharsets.UTF_8)
         val flags = Base64.DEFAULT or Base64.NO_WRAP
         return Base64.encodeToString(jsonBytes, flags)
+    }
+
+    // TODO: consider renaming this method; "match" sounds like it returns a boolean
+    fun match(
+        intent: Intent,
+        requestCode: Int,
+    ): BrowserSwitchFinishResult {
+        if (requestCode != originalOptions.requestCode) {
+            return BrowserSwitchFinishResult.RequestCodeDoesNotMatch
+        }
+
+        val deepLinkUri = intent.data
+        if (deepLinkUri == null) {
+            return BrowserSwitchFinishResult.DeepLinkNotPresent
+        }
+
+        val deepLinkScheme = deepLinkUri.scheme.orEmpty()
+        val isMatchingDeepLink =
+            deepLinkScheme.equals(originalOptions.returnUrlScheme, ignoreCase = true)
+        return if (isMatchingDeepLink) {
+            BrowserSwitchFinishResult.Success(
+                returnUrl = deepLinkUri,
+                requestCode = originalOptions.requestCode,
+                requestUrl = originalOptions.targetUri,
+                requestMetadata = originalOptions.metadata
+            )
+        } else {
+            BrowserSwitchFinishResult.DeepLinkDoesNotMatch
+        }
     }
 
     companion object {
